updates to build template [read]
patchcreator requires an exec to be in the command, which is copied to out. decided to use the readme instead of hunting for exec names and ensuring the right binary to be given

patchcreator also forces to its own gen folder, so I go up one dir if patchcreator is used. this wasnt a problem for rb3dx wii because oh how out_dir is built

we may want to think more on the concept of rebuild files.
in rb3dx you are supposed to copy the wii game files to the platform folder. Which means the xbox and ps3 folders have the dx binary, but the wii folder has the vanilla binary. not great for expansion, works for rb3
So instead I have a rebuild_files folder in dependencies with strictly dx files, that gets copied to the output instead of from the platform folder.

copy_vanilla_arks function added for patchcreator builds. i was on the fence if this should just be the default for patchcreator builds idk. because we give the user an out dir without the other ark parts, so they will have to combine them themselves. as i type this i am realizing that may be a problem for CI... welp

happy for feedback