additional template adjustments
This removes fluff options and removes the wii patchcreator edge case.

So in rb3dx for instance, how we currently have it, the "platform" folder holds the dx binaries and junk for ps3 and xbox, but the user's wii vanilla files, and that's a bit weird to expand upon
So now i am moving everything dx is providing to go in /dependencies/platform_files
This would free up /platform/ to be solely for the users vanilla files regardless of platform for use with patchcreator.

removes the vanilla ark copy function, its better to just give the users the modified files only

accounts for gen folder capitalization, ps2 is always uppercase